+++ midix/src/lib.rs
@@ -46,7 +46,7 @@ pub mod prelude {
     pub use crate::message::{controller::*, key::*, pitch_bend::*, program::*, velocity::*, *};
     pub use crate::parser::{
         reader::{ReadResult, Reader, ReaderState},
-        types::*,
+        *,
     };
 
     pub(crate) use crate::parser::reader::{inv_data, inv_input, unexp_eof};
