--- midix/src/parser/types/mod.rs
@@ -1,10 +1,0 @@
-mod header_chunk;
-pub use header_chunk::*;
-mod track_chunk;
-pub use track_chunk::*;
-mod track_event;
-pub use track_event::*;
-mod track_message;
-pub use track_message::*;
-mod event;
-pub use event::*;
